+++ test/fixtures/lib/stub-auth-token-scheme.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var Boom = require('boom');
+var Boom = require('@hapi/boom');
 
 const register = function (server, options) {
     server.auth.scheme('stub-auth-token', function (server, options) {
@@ -31,4 +31,15 @@ const register = function (server, options) {
     });
 };
 
-module.exports = { register, name: 'stub-auth-token' };
+const buildValidateFunc = function (allowedToken) {
+    return async function (token) {
+
+        if (token === allowedToken) {
+            return { credentials: { scope: [ 'api1:read' ] }, artifacts: { }};
+        }
+
+        return {};
+    }
+};
+
+module.exports = { register, name: 'stub-auth-token', buildValidateFunc};